requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
Pillow>=10.0.0
tqdm>=4.66.0
yaml>=0.2.5
PyYAML>=6.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
matplotlib>=3.8.0
scikit-learn>=1.5.0
